+++ tencentcloud/ocr/v20181119/models.go
@@ -1272,7 +1272,8 @@ type DriverLicenseOCRResponseParams struct {
 	// 有效期开始时间（YYYY-MM-DD）
 	StartDate *string `json:"StartDate,omitempty" name:"StartDate"`
 
-	// 有效期截止时间（YYYY-MM-DD）
+	// 有效期截止时间（新版驾驶证返回 YYYY-MM-DD，
+	// 老版驾驶证返回有效期限 X年）
 	EndDate *string `json:"EndDate,omitempty" name:"EndDate"`
 
 	// 证号