+++ CHANGELOG.md
@@ -1,3 +1,297 @@
+# Release v1.0.620
+
+## 文件存储(cfs) 版本：2019-07-19
+
+### 第 15 次发布
+
+发布时间：2023-03-20 01:11:04
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增数据结构：
+
+* [TieringDetailInfo](https://cloud.tencent.com/document/api/582/38175#TieringDetailInfo)
+
+修改数据结构：
+
+* [FileSystemInfo](https://cloud.tencent.com/document/api/582/38175#FileSystemInfo)
+
+	* 新增成员：TieringState, TieringDetail
+
+
+
+
+## TDSQL-C MySQL 版(cynosdb) 版本：2019-01-07
+
+### 第 49 次发布
+
+发布时间：2023-03-20 01:19:56
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [AddInstances](https://cloud.tencent.com/document/api/1003/48088)
+
+	* 新增入参：SecurityGroupIds
+
+
+
+
+## T-Sec-堡垒机（BH）(dasb) 版本：2019-10-18
+
+### 第 8 次发布
+
+发布时间：2023-03-20 01:20:40
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [Resource](https://cloud.tencent.com/document/api/1025/74416#Resource)
+
+	* 新增成员：LogDeliveryArgs
+
+
+
+
+## 低代码互动课堂(lcic) 版本：2022-08-17
+
+### 第 17 次发布
+
+发布时间：2023-03-20 01:37:28
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [GetRoomEvent](https://cloud.tencent.com/document/api/1639/88910)
+
+新增数据结构：
+
+* [EventDataInfo](https://cloud.tencent.com/document/api/1639/81423#EventDataInfo)
+* [EventInfo](https://cloud.tencent.com/document/api/1639/81423#EventInfo)
+
+
+
+## 云直播CSS(live) 版本：2018-08-01
+
+### 第 97 次发布
+
+发布时间：2023-03-20 01:38:41
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DescribeLogDownloadList](https://cloud.tencent.com/document/api/267/34084)
+
+	* 新增入参：IsFastLive
+
+
+
+
+## 云数据库 MariaDB(mariadb) 版本：2017-03-12
+
+### 第 51 次发布
+
+发布时间：2023-03-20 01:40:40
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [DescribeBackupFiles](https://cloud.tencent.com/document/api/237/88911)
+
+新增数据结构：
+
+* [InstanceBackupFileItem](https://cloud.tencent.com/document/api/237/16191#InstanceBackupFileItem)
+
+
+
+## 腾讯云可观测平台(monitor) 版本：2018-07-24
+
+### 第 80 次发布
+
+发布时间：2023-03-20 01:42:06
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateGrafanaInstance](https://cloud.tencent.com/document/api/248/79219)
+
+	* <font color="#dd0000">**修改入参**：</font>GrafanaInitPassword
+
+* [CreatePrometheusRecordRuleYaml](https://cloud.tencent.com/document/api/248/85707)
+
+	* 新增入参：Name
+
+
+
+
+## 媒体处理(mps) 版本：2019-06-12
+
+### 第 54 次发布
+
+发布时间：2023-03-20 01:43:46
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [OverrideTranscodeParameter](https://cloud.tencent.com/document/api/862/37615#OverrideTranscodeParameter)
+
+	* 新增成员：AddonAudioStream
+
+
+
+
+## 集团账号管理(organization) 版本：2021-03-31
+
+### 第 16 次发布
+
+发布时间：2023-03-20 01:46:48
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [OrgMemberAuthIdentity](https://cloud.tencent.com/document/api/850/67060#OrgMemberAuthIdentity)
+
+	* 新增成员：IdentityType
+
+
+
+
+## 集团账号管理(organization) 版本：2018-12-25
+
+
+
+## SSL 证书(ssl) 版本：2019-12-05
+
+### 第 27 次发布
+
+发布时间：2023-03-20 01:52:16
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [UploadCertificate](https://cloud.tencent.com/document/api/400/41665)
+
+	* 新增入参：Repeatable
+
+	* 新增出参：RepeatCertId
+
+
+
+
+## 容器镜像服务(tcr) 版本：2019-09-24
+
+### 第 53 次发布
+
+发布时间：2023-03-20 01:55:07
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateInstance](https://cloud.tencent.com/document/api/1141/41572)
+
+	* 新增入参：EnableCosMAZ
+
+
+
+
+## 容器安全服务(tcss) 版本：2020-11-01
+
+### 第 36 次发布
+
+发布时间：2023-03-20 01:55:56
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateVirusScanTask](https://cloud.tencent.com/document/api/1285/65584)
+
+	* 新增入参：ScanPathMode
+
+* [DescribeVirusDetail](https://cloud.tencent.com/document/api/1285/65583)
+
+	* 新增出参：FileAccessTime, FileModifyTime
+
+* [DescribeVirusMonitorSetting](https://cloud.tencent.com/document/api/1285/65581)
+
+	* 新增出参：ScanPathMode
+
+* [DescribeVirusScanSetting](https://cloud.tencent.com/document/api/1285/65580)
+
+	* 新增出参：ScanPathMode
+
+* [ModifyVirusMonitorSetting](https://cloud.tencent.com/document/api/1285/65573)
+
+	* 新增入参：ScanPathMode
+
+* [ModifyVirusScanSetting](https://cloud.tencent.com/document/api/1285/65572)
+
+	* 新增入参：ScanPathMode
+
+
+
+
+## 边缘安全加速平台(teo) 版本：2022-09-01
+
+### 第 36 次发布
+
+发布时间：2023-03-20 02:00:54
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增数据结构：
+
+* [ImageOptimize](https://cloud.tencent.com/document/api/1552/80721#ImageOptimize)
+
+修改数据结构：
+
+* [ZoneSetting](https://cloud.tencent.com/document/api/1552/80721#ZoneSetting)
+
+	* 新增成员：ImageOptimize
+
+
+
+
+## 边缘安全加速平台(teo) 版本：2022-01-06
+
+
+
 # Release v1.0.619
 
 ## 云数据仓库 ClickHouse(cdwch) 版本：2020-09-15